+++ resources/examples/manager/BasicAssetLibrary/tests/fixtures.py
@@ -41,6 +41,7 @@ identifier = "org.openassetio.examples.manager.bal"
 
 valid_ref = "bal:///references/can/contain/🐠"
 non_ref = "not a Ŕeference"
+malformed_ref = "bal:///"
 
 an_existing_entity_name = next(iter(test_library["entities"].keys()))
 
@@ -52,7 +53,8 @@ fixtures = {
     },
     "shared": {
         "a_valid_reference": valid_ref,
-        "a_malformed_reference": non_ref,
+        "an_invalid_reference": non_ref,
+        "a_malformed_reference": malformed_ref,
     },
     "Test_identifier": {
         "test_matches_fixture": {
@@ -97,8 +99,8 @@ fixtures = {
             "a_reference_to_a_missing_entity": "bal:///missing_entity",
             "the_error_string_for_a_reference_to_a_missing_entity": (
                 "Entity 'bal:///missing_entity' not found"),
-            "a_malformed_entity_reference": f"bal:///",
-            "the_error_string_for_a_malformed_entity_reference": (
+            "a_malformed_reference": malformed_ref,
+            "the_error_string_for_a_malformed_reference": (
                 f"Missing entity name in path component")
         }
     }